+++ ui/server/graphql/resolvers/group/index.ts
@@ -1,0 +1,1 @@
+export * as queries from "./queries";
