+++ ui/server/graphql/resolvers/index.ts
@@ -1,4 +1,3 @@
-import SeededShuffle from "seededshuffle";
 import slugify from "../../utils/slugify";
 //import liveUpdate from "../../services/liveUpdate.service";
 import prisma from "../../prisma";
@@ -44,31 +43,15 @@ import { updateFundedPercentage } from "../resolvers/helpers";
 import {
   createGroupInvitationLink,
   deleteGroupInvitationLink,
-  groupInvitationLink,
-} from "./group";
+} from "./group/mutations";
+
+//queries
+import { queries as userQueries } from "./user";
+import { queries as groupQueries } from "./group";
+import { queries as roundQueries } from "./round";
+import { queries as bucketQueries } from "./bucket";
 
 const { groupHasDiscourse, generateComment } = subscribers;
-
-const isRootAdmin = (parent, args, { user }) => {
-  // TODO: this is old code that doesn't really work right now
-  return user && user.isRootAdmin
-    ? skip
-    : new Error("You need to be root admin");
-};
-
-const isMemberOfGroup = async (parent, { groupId }, { user }) => {
-  if (!user) throw new Error("You need to be logged in");
-
-  const currentGroupMember = await prisma.groupMember.findUnique({
-    where: {
-      groupId_userId: { groupId: groupId, userId: user.id },
-    },
-  });
-
-  if (!currentGroupMember)
-    throw new Error("You need to be a member of that group");
-  return skip;
-};
 
 const isCollMember = async (parent, { roundId, bucketId }, { user }) => {
   if (!user) throw new Error("You need to be logged in");
@@ -89,32 +72,6 @@ const isCollMember = async (parent, { roundId, bucketId }, { user }) => {
     throw new Error("Round member has not accepted the invitation");
   }
 
-  return skip;
-};
-
-const isCollMemberOrGroupAdmin = async (parent, { roundId }, { user }) => {
-  if (!user) throw new Error("You need to be logged in");
-
-  const roundMember = await getRoundMember({
-    userId: user.id,
-    roundId,
-  });
-
-  let groupMember = null;
-  if (!roundMember) {
-    const group = await prisma.group.findFirst({
-      where: { rounds: { some: { id: roundId } } },
-    });
-    groupMember = await getGroupMember({
-      userId: user.id,
-      groupId: group.id,
-    });
-  }
-
-  if (!(roundMember?.isApproved || groupMember?.isAdmin))
-    throw new Error(
-      "You need to be an approved participant in this round or a group admin to view round participants"
-    );
   return skip;
 };
 
@@ -177,436 +134,13 @@ const isBucketCocreatorOrCollAdminOrMod = async (
 
 const resolvers = {
   Query: {
-    currentUser: async (parent, args, { user }) => {
-      return user
-        ? await prisma.user.findUnique({ where: { id: user.id } })
-        : null;
-    },
-    user: async (parent, { userId }) => {
-      // we let the resolvers grab any extra requested fields, so we don't accidentally leak e.g. emails
-      return prisma.user.findUnique({
-        where: { id: userId },
-        select: { id: true },
-      });
-    },
-    group: async (parent, { groupSlug }, { user }) => {
-      if (!groupSlug) return null;
-      if (process.env.SINGLE_GROUP_MODE !== "true" && groupSlug == "c")
-        return null;
-      return getGroup({ groupSlug, user });
-    },
-    groups: combineResolvers(isRootAdmin, async (parent, args) => {
-      return prisma.group.findMany();
-    }),
-    rounds: async (parent, { limit, groupSlug }, { user }) => {
-      if (!groupSlug) return null;
+    ...userQueries,
+    ...groupQueries,
+    ...roundQueries,
+    ...bucketQueries,
 
-      const currentGroupMember = user
-        ? await prisma.groupMember.findFirst({
-            where: {
-              group: { slug: groupSlug },
-              userId: user.id,
-            },
-          })
-        : null;
-
-      // if admin show all rounds (current or archived)
-      if (currentGroupMember?.isAdmin) {
-        return prisma.round.findMany({
-          where: { group: { slug: groupSlug }, deleted: { not: true } },
-          take: limit,
-        });
-      }
-
-      const allRounds = await prisma.round.findMany({
-        where: {
-          group: { slug: groupSlug },
-          archived: { not: true },
-          deleted: { not: true },
-        },
-        take: limit,
-      });
-
-      // filter away colls the current user shouldn't be able to view
-      return (
-        await Promise.all(
-          allRounds.map(async (coll) =>
-            (await canViewRound({ round: coll, user })) ? coll : undefined
-          )
-        )
-      ).filter(Boolean);
-    },
-    round: async (parent, { groupSlug, roundSlug }, { user }) => {
-      if (!roundSlug) return null;
-
-      const round = await prisma.round.findFirst({
-        where: {
-          slug: roundSlug,
-          group: { slug: groupSlug ?? "c" },
-          deleted: { not: true },
-        },
-      });
-      if (!round) return null;
-
-      if (await canViewRound({ round: round, user })) {
-        return round;
-      } else {
-        return null;
-      }
-    },
-
-    groupInvitationLink,
-
-    invitationLink: async (parent, { roundId }, { user }) => {
-      const isAdmin =
-        !!user &&
-        isCollAdmin({
-          userId: user.id,
-          roundId,
-        });
-
-      if (!isAdmin) {
-        throw new Error("You need to be admin to fetch invitation link");
-      }
-
-      const round = await prisma.round.findFirst({
-        where: {
-          id: roundId,
-        },
-      });
-      return {
-        link:
-          round.inviteNonce !== null
-            ? appLink("/invite/" + sign({ nonce: round.inviteNonce, roundId }))
-            : null,
-      };
-    },
-    contributionsPage: combineResolvers(
-      isCollMemberOrGroupAdmin,
-      async (parent, { roundId, offset, limit }) => {
-        // const contributionsWithExtra = [
-        //   ...(await Contribution.find({ roundId }, null, {
-        //     skip: offset,
-        //     limit: limit + 1,
-        //   }).sort({
-        //     createdAt: -1,
-        //   })),
-        // ];
-
-        const contributionsWithExtra = await prisma.contribution.findMany({
-          where: { roundId },
-          take: limit,
-          skip: offset,
-          orderBy: {
-            createdAt: "desc",
-          },
-        });
-
-        return {
-          moreExist: contributionsWithExtra.length > limit,
-          contributions: contributionsWithExtra.slice(0, limit),
-        };
-      }
-    ),
-    roundTransactions: combineResolvers(
-      isCollMember,
-      async (parent, { roundId, offset, limit }) => {
-        const transactions: [RoundTransaction] = await prisma.$queryRaw`
-          (
-            SELECT 
-              "id", 
-              "collectionMemberId" as "roundMemberId", 
-              null as "allocatedById", 
-              "amount",
-              "bucketId",
-              "amountBefore", 
-              null as "allocationType",
-              'CONTRIBUTION' as "transactionType",
-              "createdAt"
-            FROM "Contribution" where "collectionId" = ${roundId}
-            
-            UNION ALL
-            
-            SELECT 
-              "id", 
-              "collectionMemberId" as "roundMemberId", 
-              "allocatedById", 
-              "amount",
-              null as "bucketId",
-              "amountBefore", 
-              "allocationType",
-              'ALLOCATION' as "transactionType",
-              "createdAt"
-            FROM "Allocation" where "collectionId" = ${roundId}
-          ) ORDER BY "createdAt" DESC LIMIT ${limit} OFFSET ${offset};
-        `;
-
-        transactions.forEach(
-          (transaction) =>
-            (transaction.createdAt = new Date(transaction.createdAt))
-        );
-
-        return {
-          moreExist: transactions.length > limit,
-          transactions: transactions.slice(0, limit),
-        };
-      }
-    ),
-    bucket: async (parent, { id }) => {
-      if (!id) return null;
-      const bucket = await prisma.bucket.findUnique({ where: { id } });
-      if (!bucket || bucket.deleted) return null;
-      return bucket;
-    },
     languageProgressPage: async () => {
       return getLanguageProgress();
-    },
-    bucketsPage: async (
-      parent,
-      {
-        roundSlug,
-        groupSlug,
-        textSearchTerm,
-        tag: tagValue,
-        offset = 0,
-        limit,
-        status,
-        orderBy,
-        orderDir,
-      },
-      { user }
-    ) => {
-      const currentMember = await prisma.roundMember.findFirst({
-        where: {
-          userId: user?.id ?? "undefined",
-          round: { slug: roundSlug, group: { slug: groupSlug ?? "c" } },
-        },
-      });
-
-      const isAdminOrGuide =
-        currentMember && (currentMember.isAdmin || currentMember.isModerator);
-
-      const statusFilter = status.map(statusTypeToQuery).filter((s) => s);
-
-      const buckets = await prisma.bucket.findMany({
-        where: {
-          round: { slug: roundSlug, group: { slug: groupSlug ?? "c" } },
-          deleted: { not: true },
-          OR: statusFilter,
-          ...(textSearchTerm && { title: { search: textSearchTerm } }),
-          ...(tagValue && {
-            tags: { some: { value: tagValue } },
-          }),
-          ...(!isAdminOrGuide &&
-            (currentMember
-              ? {
-                  OR: [
-                    { publishedAt: { not: null } },
-                    { cocreators: { some: { id: currentMember.id } } },
-                  ],
-                }
-              : { publishedAt: { not: null } })),
-        },
-        ...(orderBy && { orderBy: { [orderBy]: orderDir } }),
-      });
-
-      const todaySeed = dayjs().format("YYYY-MM-DD");
-
-      let shuffledBuckets = buckets;
-      if (!orderBy) {
-        SeededShuffle.shuffle(buckets, user ? user.id + todaySeed : todaySeed);
-      } else if (
-        orderBy === "percentageFunded" ||
-        orderBy === "contributionsCount"
-      ) {
-        // Move nulls to last manually. This feature is added to prisma but not pushed to release yet
-        // https://github.com/prisma/prisma-engines/pull/3036
-
-        const index = shuffledBuckets.findIndex((b) => b[orderBy] !== null);
-        const length = shuffledBuckets.length;
-
-        shuffledBuckets = shuffledBuckets
-          .splice(index, length - index)
-          .concat(shuffledBuckets);
-      }
-
-      return {
-        moreExist: shuffledBuckets.length > limit + offset,
-        buckets: shuffledBuckets.slice(offset, limit + offset),
-      };
-    },
-    groupMembersPage: combineResolvers(
-      isGroupAdmin,
-      async (
-        parent,
-        { offset = 0, limit, groupId, search, isApproved },
-        { user }
-      ) => {
-        const groupMembersWithExtra = await prisma.groupMember.findMany({
-          where: {
-            groupId: groupId,
-            isApproved,
-            ...(search && {
-              user: {
-                OR: [
-                  { username: { contains: search, mode: "insensitive" } },
-                  { name: { contains: search, mode: "insensitive" } },
-                  { email: { contains: search, mode: "insensitive" } },
-                ],
-              },
-            }),
-          },
-          skip: offset || 0,
-          take: (limit || 1e3) + 1,
-        });
-
-        return {
-          moreExist: groupMembersWithExtra.length > limit,
-          groupMembers: groupMembersWithExtra.slice(0, limit),
-        };
-      }
-    ),
-    members: combineResolvers(
-      isCollMemberOrGroupAdmin,
-      async (parent, { roundId, isApproved }) => {
-        return await prisma.roundMember.findMany({
-          where: {
-            roundId,
-            isApproved,
-            ...(!isApproved && { isRemoved: false }),
-          },
-        });
-      }
-    ),
-    membersPage: combineResolvers(
-      isCollMemberOrGroupAdmin,
-      async (
-        parent,
-        { roundId, isApproved, search, offset = 0, limit = 10 },
-        { user }
-      ) => {
-        const isAdmin = await isCollAdmin({
-          userId: user.id,
-          roundId,
-        });
-        if (!isAdmin && !isApproved) return null;
-
-        const insensitiveSearch: { contains: string; mode: "insensitive" } = {
-          contains: search,
-          mode: "insensitive",
-        };
-
-        const roundMembersWithExtra = await prisma.roundMember.findMany({
-          where: {
-            roundId,
-            isApproved,
-            ...(!isApproved && { isRemoved: false }),
-            ...(search && {
-              user: {
-                OR: [
-                  { username: insensitiveSearch },
-                  { name: insensitiveSearch },
-                  ...(isAdmin ? [{ email: insensitiveSearch }] : []),
-                ],
-              },
-            }),
-            ...(!isAdmin && { hasJoined: true }),
-          },
-          take: limit + 1,
-          skip: offset,
-          ...(search && { include: { user: true } }),
-        });
-
-        return {
-          moreExist: roundMembersWithExtra.length > limit,
-          members: roundMembersWithExtra.slice(0, limit),
-        };
-      }
-    ),
-    categories: async (parent, { groupId }) => {
-      const group = await prisma.group.findUnique({
-        where: { id: groupId },
-        include: { discourse: true },
-      });
-
-      if (!group.discourse) {
-        return [];
-      }
-
-      // TODO: permission check here?
-
-      const categories = await discourse(group.discourse).categories.getAll();
-
-      return categories;
-    },
-    commentSet: async (
-      parent,
-      { bucketId, from = 0, limit = 30, order = "desc" }
-    ) => {
-      const bucket = await prisma.bucket.findUnique({
-        where: { id: bucketId },
-        include: {
-          comments: true,
-          round: {
-            include: { group: { include: { discourse: true } } },
-          },
-        },
-      });
-      // const bucket = await Bucket.findOne({ _id: bucketId });
-
-      let comments;
-      const group = bucket.round.group;
-
-      if (groupHasDiscourse(group)) {
-        const topic = await discourse(group.discourse).posts.get(
-          bucket.discourseTopicId
-        );
-
-        comments = await Promise.all(
-          topic.post_stream.posts
-            .filter((post) => post.post_number > 1)
-            .filter((post) => !post.user_deleted)
-            // filter out empty system comments, e.g. when a thread is moved
-            .filter(
-              (comment) =>
-                !(comment.username === "system" && comment.raw === "")
-            )
-            .reverse()
-            .map(async (post) => {
-              const author = await prisma.roundMember.findFirst({
-                where: {
-                  roundId: bucket.roundId,
-                  user: {
-                    groupMemberships: {
-                      some: {
-                        discourseUsername: post.username,
-                        groupId: group.id,
-                      },
-                    },
-                  },
-                },
-              });
-
-              return generateComment(post, author);
-            })
-        );
-      } else {
-        comments = await prisma.comment.findMany({
-          where: { bucketId: bucketId },
-          orderBy: { createdAt: "desc" },
-        });
-      }
-
-      let shown = comments.slice(0, from + limit);
-
-      if (order === "desc") {
-        shown = shown.reverse();
-      }
-
-      return {
-        total: comments.length,
-        comments: shown,
-      };
     },
   },
   Mutation: {
