+++ ui/server/graphql/resolvers/round/index.ts
@@ -1,0 +1,1 @@
+export * as queries from "./queries";
