+++ ui/server/graphql/resolvers/user/index.ts
@@ -1,0 +1,1 @@
+export * as queries from "./queries";
